Split knobMaterial.contract.test.ts into focused test modules
The 1920-line contract test file is decomposed by natural test seams
(WGSL shader snapshot, three Canvas2D draw-call snapshots by knob
value, and geometry/derivation checks) into separate files under
src/components/__tests__/knobMaterial/, with shared test fixtures
extracted to shared.ts. No behavior or assertions changed; each file
is now well under the 700-line target.
