+++ src/components/__tests__/knobMaterial/wgsl.contract.test.ts
@@ -1,0 +1,187 @@
+import { describe, expect, it } from 'vitest';
+
+import { buildKnobShaderCode } from '../../KnobGPUContext';
+import { KNOB_MATERIAL } from '../../knobMaterial';
+
+describe('holographic knob derivation contract', () => {
+    it('snapshots WGSL derived from KNOB_MATERIAL', () => {
+        const shader = buildKnobShaderCode(KNOB_MATERIAL);
+        expect(shader).toMatchInlineSnapshot(`
+          "
+              struct Uniforms {
+                time: f32,
+                value: f32,
+                resolution: vec2f,
+              };
+              @group(0) @binding(0) var<uniform> u: Uniforms;
+
+              struct VertexOutput {
+                @builtin(position) position: vec4f,
+                @location(0) uv: vec2f,
+              };
+
+              @vertex
+              fn vs_main(@builtin(vertex_index) vIdx: u32) -> VertexOutput {
+                // Full screen triangle
+                var pos = array<vec2f, 3>(
+                  vec2f(-1.0, -1.0),
+                  vec2f(3.0, -1.0),
+                  vec2f(-1.0, 3.0)
+                );
+                var output: VertexOutput;
+                output.position = vec4f(pos[vIdx], 0.0, 1.0);
+                output.uv = pos[vIdx]; // UVs are -1 to 1 effectively for the quad
+                return output;
+              }
+
+              // Helper: Circle SDF
+              fn sdCircle(p: vec2f, r: f32) -> f32 {
+                  return length(p) - r;
+              }
+              
+              // Helper: Rotation Matrix
+              fn rotate(angle: f32) -> mat2x2f {
+                  let c = cos(angle);
+                  let s = sin(angle);
+                  return mat2x2f(c, -s, s, c);
+              }
+
+              @fragment
+              fn fs_main(in: VertexOutput) -> @location(0) vec4f {
+                var uv = in.uv;
+                // Center UVs properly assuming canvas is square-ish
+                // Standardize coordinate system to -1.0 to 1.0
+                
+                let len = length(uv);
+                let angle = atan2(uv.y, uv.x);
+                
+                // Base Color (Cyan/Teal Hologram)
+                var color = vec3f(0.0000, 0.8980, 1.0000);
+                var alpha = 0.0;
+
+                // 1. Outer Data Ring
+                // Rotating dashed ring
+                let rot_uv = rotate(u.time * 0.2) * uv;
+                let ring_dist = abs(length(rot_uv) - 0.5500);
+                let dash = sin(atan2(rot_uv.y, rot_uv.x) * 20.0);
+                if (ring_dist < 0.02 && dash > 0.5) {
+                    alpha += 0.6 * smoothstep(0.02, 0.0, ring_dist);
+                }
+
+                // 2. Value Arc (The "Level" Indicator)
+                // Map value 0..1 to angle range [-max_angle, +max_angle]
+                // Knob starts at bottom-left, sweeps clockwise to bottom-right.
+                let max_angle = 2.3562;
+                let val_mapped = mix(-max_angle, max_angle, u.value);
+
+                // Needle direction: rotate standard "up" (+Y) by val_mapped
+                let needle_vec = vec2f(sin(val_mapped), cos(val_mapped));
+
+                // Value arc: light the inner ring where the pixel angle <= val_mapped
+                // atan2 gives [-π, π]; we shift so 0 = straight up (matches needle_vec above)
+                let pixel_angle = atan2(uv.x, uv.y); // note: swapped args for "up = 0" convention
+                let arc_radius = 0.4200;
+                let arc_dist = abs(length(uv) - arc_radius);
+                // Only draw arc within the ±max_angle sweep and up to the current value
+                if (arc_dist < 0.03 && pixel_angle >= -max_angle && pixel_angle <= val_mapped) {
+                    let arc_brightness = smoothstep(0.03, 0.0, arc_dist);
+                    // Color shifts from teal at min to bright cyan at current value
+                    let arc_t = (pixel_angle + max_angle) / (val_mapped + max_angle + 0.001);
+                    color = mix(vec3f(0.0000, 0.6000, 0.5000), vec3f(0.2000, 1.0000, 0.8000), arc_t);
+                    alpha += arc_brightness * 0.85;
+                }
+
+                // 2b. Scale ticks along the value arc (hardware-style position reference)
+                let tick_angles = array<f32, 7>(-2.356194, -1.570796, -0.785398, 0.000000, 0.785398, 1.570796, 2.356194);
+                let tick_major = array<f32, 7>(1, 0, 0, 1, 0, 0, 1);
+                let tick_count = 7u;
+                for (var ti = 0u; ti < tick_count; ti++) {
+                    let ta = tick_angles[ti];
+                    let angle_diff = abs(pixel_angle - ta);
+                    let half_len = mix(0.0300, 0.0450, tick_major[ti]);
+                    let tick_inner = arc_radius - half_len;
+                    let tick_outer = arc_radius + half_len;
+                    let r = length(uv);
+                    let tick_color = mix(vec3f(0.0000, 0.7500, 0.8500), vec3f(0.2000, 1.0000, 1.0000), tick_major[ti]);
+                    if (r >= tick_inner && r <= tick_outer && angle_diff < 0.035 && pixel_angle >= -max_angle && pixel_angle <= max_angle) {
+                        let tick_brightness = smoothstep(0.035, 0.0, angle_diff) * mix(0.55, 0.95, tick_major[ti]);
+                        color = tick_color;
+                        alpha += tick_brightness;
+                    }
+                }
+
+                // 2c. Detent dimples (recessed hardware notches at major landmarks)
+                let detent_angles = array<f32, 3>(-2.356194, 0.000000, 2.356194);
+                let detent_count = 3u;
+                let detent_shadow = vec3f(0.0000, 0.1500, 0.2000);
+                for (var di = 0u; di < detent_count; di++) {
+                    let da = detent_angles[di];
+                    let detent_diff = abs(pixel_angle - da);
+                    let r = length(uv);
+                    let notch_inner = arc_radius - 0.0225;
+                    let notch_outer = arc_radius + 0.0225 * 0.65;
+                    if (detent_diff < 0.0550 && r >= notch_inner && r <= notch_outer && pixel_angle >= -max_angle && pixel_angle <= max_angle) {
+                        let dimple = smoothstep(0.0550, 0.0, detent_diff)
+                            * smoothstep(notch_outer, arc_radius, r)
+                            * smoothstep(notch_inner, arc_radius, r);
+                        color = mix(color, detent_shadow, dimple * 0.85);
+                        alpha += dimple * 0.55;
+                    }
+                }
+
+                // 3. Holographic Scanlines
+                let scanline = sin(uv.y * 150.0 + u.time * 10.0) * 0.5 + 0.5;
+
+                // 4. Central Glow / "Projector" beam
+                let beam = smoothstep(0.6, 0.0, len);
+
+                // Compose the "Ghost" Knob
+                var final_c = vec3f(0.0);
+
+                // Main Circle Body
+                let circle_edge = 1.0 - smoothstep(0.48, 0.5, len);
+                let inner_glow = smoothstep(0.0, 0.5, len);
+                alpha += circle_edge * 0.2 * scanline; // Background body
+
+                // The Needle — hardware-lit variant or legacy glow
+                let proj = dot(uv, needle_vec);
+                let perp = length(uv - needle_vec * proj);
+                let needle_len = 0.5000;
+                
+                let needle_half_w = 0.018;
+                if (proj > 0.02 && proj < needle_len && perp < needle_half_w) {
+                    let along = proj / needle_len;
+                    let body = mix(vec3f(0.0500, 0.0800, 0.1200), vec3f(0.8200, 0.8600, 0.9000), smoothstep(0.0, 1.0, along));
+                    let specular = exp(-pow((along - 0.68) / 0.11, 2.0))
+                        * exp(-pow(perp / (needle_half_w * 0.35), 2.0));
+                    color = mix(body, vec3f(1.0000, 1.0000, 1.0000), specular * 0.95);
+                    alpha += smoothstep(needle_half_w, 0.0, perp) * 0.95;
+                }
+                // Cast shadow opposite the needle on the knob face
+                let shadow_vec = -needle_vec;
+                let shadow_proj = dot(uv, shadow_vec);
+                let shadow_perp = length(uv - shadow_vec * shadow_proj);
+                if (shadow_proj > 0.02 && shadow_proj < 0.14 && shadow_perp < 0.07) {
+                    let shadow_a = smoothstep(0.14, 0.02, shadow_proj) * smoothstep(0.07, 0.0, shadow_perp) * 0.35;
+                    color = mix(color, vec3f(0.0500, 0.0800, 0.1200), shadow_a);
+                    alpha += shadow_a * 0.4;
+                }
+                
+
+                // 5. Fresnel / Glitch Effect
+                let glitch = step(0.98, sin(u.time * 20.0 + uv.y * 10.0));
+                if (glitch > 0.5) {
+                    uv.x += 0.05;
+                    alpha += 0.2;
+                }
+
+                // Vignette / Falloff at edges of canvas
+                alpha *= smoothstep(0.8, 0.6, len);
+
+                return vec4f(color * alpha * 1.5000, alpha);
+              }
+          "
+        `);
+        expect(shader).toMatchSnapshot();
+    });
+});
